+++ rustls/src/msgs/handshake.rs
@@ -10,8 +10,8 @@ use core::{fmt, iter};
 use pki_types::{CertificateDer, DnsName};
 
 use crate::crypto::cipher::Payload;
-use crate::crypto::ffdhe_groups::FfdheGroup;
 use crate::crypto::hpke::{HpkeKem, HpkeSymmetricCipherSuite};
+use crate::crypto::kx::ffdhe::FfdheGroup;
 use crate::crypto::kx::{ActiveKeyExchange, KeyExchangeAlgorithm};
 use crate::crypto::{
     CipherSuite, GetRandomFailed, NamedGroup, SecureRandom, SelectedCredential, SignatureScheme,
